+++ Data/Repositories/RollRepository.cs
@@ -18,7 +18,7 @@ namespace DAO.Repositories
         {
             _connectionString = connectionString;
         }
-        public bool Delete(int id)
+        public bool Delete(long id)
         {
             bool success = false;
             try
@@ -50,7 +50,7 @@ namespace DAO.Repositories
             catch (Exception ex) { Console.WriteLine($"Exception in {GetType()}.Add:" + ex.Message); }
             return success;
         }
-        public RollEntity? GetById(int id)
+        public RollEntity? GetById(long id)
         {
             try
             {
