+++ Data/Repositories/UserDataRepository.cs
@@ -19,7 +19,7 @@ namespace DAO.Repositories
         {
             _connectionString = connectionString;
         }
-        public bool Delete(int id)
+        public bool Delete(long id)
         {
             bool success = false;
             try
@@ -56,7 +56,7 @@ namespace DAO.Repositories
 
             return success;
         }
-        public UserDataEntity? GetById(int id)
+        public UserDataEntity? GetById(long id)
         {
             try
             {
